fix(staging): fixing for test deployment on railway
