+++ package.json
@@ -72,5 +72,9 @@
     "name": "Bankole Esan",
     "email": "techybanky@gmail.com",
     "url": "https://banky.studio"
+  },
+  "engines": {
+    "npm": ">=8.0.0 <9.0.0",
+    "node": ">=16.0.0 <17.0.0"
   }
 }
