+++ pages/dashboard/admin.vue
@@ -5,7 +5,9 @@
 </template>
 
 <script>
-export default {}
+export default {
+  layout: 'admin',
+}
 </script>
 
 <style></style>
