+++ pages/dashboard/index/index.vue
@@ -65,7 +65,7 @@
 
 <script>
 export default {
-  layout: 'admin',
+  layout: 'app',
   middleware: ['checkAuth', 'requireUser'],
   data() {
     return {
