+++ pages/post/index.vue
@@ -7,7 +7,7 @@
 
 <script>
 export default {
-  layout: 'admin',
+  layout: 'app',
 }
 </script>
