+++ pages/profile.vue
@@ -9,7 +9,7 @@
 
 <script>
 export default {
-  layout: 'admin',
+  layout: 'app',
   scrollToTop: true,
   data() {
     return {
